week3/rag-chatbot/requirements.txt
streamlit
langchain
langchain-openai
openai
faiss-cpu
pypdf
python-dotenv
serpapi
tiktoken
